Changed header img to bbcode
The forums this script is used on, has bbcode added to use the [recheader] instead of an image. This way every time some form of leadership changes on the recruitment header, it needs to be updated in the code. Now it just needs to be updated on the forum itself.